+++ src/lib/algorithms/sequential/dinics.cpp
@@ -12,13 +12,13 @@
 void DinicsSequentialSolver::initialize(MaxFlowInstance *input){
   num_vertices = input->inputGraph.num_vertices;
   levels = new int[num_vertices];
-  flows = new float*[num_vertices];
+  flows = new int*[num_vertices];
   capacities = input->inputGraph.capacities;
   edges.clear();
 
 
   for (int i = 0; i < num_vertices; i++) {
-    flows[i] = new float[num_vertices];
+    flows[i] = new int[num_vertices];
     std::vector<int> adj;
     for (int j = 0; j < num_vertices; j++) {
       flows[i][j] = 0;
@@ -30,7 +30,7 @@ void DinicsSequentialSolver::initialize(MaxFlowInstance *input){
   }
 }
 
-float DinicsSequentialSolver::sendFlow(int current, int flow, int sink, int start[])
+int DinicsSequentialSolver::sendFlow(int current, int flow, int sink, int start[])
 {
   // Sink reached
   if (current == sink)
@@ -44,8 +44,8 @@ float DinicsSequentialSolver::sendFlow(int current, int flow, int sink, int star
     if (levels[child] == levels[current]+1 && flows[current][child] < capacities[current][child])
     {
       // find minimum flow from current to sink
-      float curr_flow = fmin(flow, capacities[current][child] - flows[current][child]);
-      float temp_flow = sendFlow(child, curr_flow, sink, start);
+      int curr_flow = fmin(flow, capacities[current][child] - flows[current][child]);
+      int temp_flow = sendFlow(child, curr_flow, sink, start);
 
       // flow is greater than zero
       if (temp_flow > 0)
